# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "McDee Platform API",
        "description": "Multifunctional Marketplace & Services API",
        "contact": {
            "email": "support@mcdee.test"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000",
            "description": "Local Server"
        }
    ],
    "paths": {
        "/api/apartment/bookings": {
            "get": {
                "tags": [
                    "Apartment Bookings"
                ],
                "summary": "Get my apartment bookings",
                "operationId": "1024bba4fdcc966175c70c8f1e093e46",
                "responses": {
                    "200": {
                        "description": "List of bookings"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Apartment Bookings"
                ],
                "summary": "Book an apartment",
                "operationId": "b9e7c7b2de839b4bcab677e31f7bba87",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "apartment_id",
                                    "check_in",
                                    "check_out"
                                ],
                                "properties": {
                                    "apartment_id": {
                                        "type": "integer"
                                    },
                                    "check_in": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "check_out": {
                                        "type": "string",
                                        "format": "date"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Booking created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/apartments": {
            "get": {
                "tags": [
                    "Service Apartments"
                ],
                "summary": "List all apartments",
                "operationId": "a8e6ba6dd64e1e9208b735934467a0be",
                "responses": {
                    "200": {
                        "description": "List of apartments"
                    }
                }
            },
            "post": {
                "tags": [
                    "Service Apartments"
                ],
                "summary": "Add new apartment (Vendor only)",
                "operationId": "c40dbd16366fb6f100e6bbf98eee2768",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "location",
                                    "price_per_night",
                                    "type"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Luxury Shortlet Apartment"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Spacious apartment with sea view."
                                    },
                                    "location": {
                                        "type": "string",
                                        "example": "Lekki Phase 1, Lagos"
                                    },
                                    "price_per_night": {
                                        "type": "number",
                                        "example": 25000
                                    },
                                    "type": {
                                        "type": "string",
                                        "enum": [
                                            "hotel",
                                            "hostel",
                                            "shortlet"
                                        ],
                                        "example": "shortlet"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "example": "image1.jpg"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Apartment created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/apartments/{id}": {
            "get": {
                "tags": [
                    "Service Apartments"
                ],
                "summary": "Get apartment details",
                "operationId": "8731d0bfce1620b0c185afb5bda9c6db",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Apartment details"
                    }
                }
            }
        },
...
        "/api/maintenance/request": {
            "post": {
                "tags": [
                    "Auto Maintenance"
                ],
                "summary": "Submit a vehicle maintenance request",
                "operationId": "a4d9533d5cca40e4148dc02ddd1af40e",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "location": {
                                        "type": "string",
                                        "example": "123 Main Street, Lagos"
                                    },
                                    "service_type": {
                                        "type": "string",
                                        "example": "Engine Repair"
                                    },
                                    "issue": {
                                        "type": "string",
                                        "example": "Car won’t start and makes noise"
                                    },
                                    "needs_towing": {
                                        "type": "boolean",
                                        "example": true
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Request created successfully"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/maintenance/my-requests": {
            "get": {
                "tags": [
                    "Auto Maintenance"
                ],
                "summary": "Get logged-in user's maintenance requests",
                "operationId": "d0d27bcd7a9c084cb5634d16873155c2",
                "responses": {
                    "200": {
                        "description": "List of maintenance requests"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/maintenance/update-status": {
            "put": {
                "tags": [
                    "Auto Maintenance"
                ],
                "summary": "Update the status of a maintenance request",
                "operationId": "24359462f23e0b7834a9d1388d1b98b0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "request_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "accepted",
                                            "completed",
                                            "cancelled"
                                        ],
                                        "example": "accepted"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    },
                    "404": {
                        "description": "Request not found"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/request": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Request a ride",
                "operationId": "28fe8f8ee86fabf1561b68865f92b14b",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "pickup_lat": {
                                        "type": "number"
                                    },
                                    "pickup_lng": {
                                        "type": "number"
                                    },
                                    "dropoff_lat": {
                                        "type": "number"
                                    },
                                    "dropoff_lng": {
                                        "type": "number"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Ride created"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/update-status": {
            "put": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Update ride status",
                "operationId": "03ac2eea29ca76ab96272c27231e3320",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ride_id": {
                                        "type": "integer"
                                    },
                                    "status": {
                                        "type": "string",
                                        "enum": [
                                            "on_trip",
                                            "completed",
                                            "cancelled"
                                        ]
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Status updated"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/history": {
            "get": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Get ride history",
                "operationId": "71dc7f2b11e22a6e1f6e601263023896",
                "responses": {
                    "200": {
                        "description": "List of rides"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/online": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Set rider online",
                "operationId": "62fabb36cf75595ba17def019828a738",
                "responses": {
                    "200": {
                        "description": "Rider is now online"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/offline": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Set rider offline",
                "operationId": "6a063441e2de6453535bf7870eaf2d12",
                "responses": {
                    "200": {
                        "description": "Rider is now offline"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/ride/rate": {
            "post": {
                "tags": [
                    "Ride Hailing"
                ],
                "summary": "Rate a completed ride",
                "operationId": "44ab5cf6e8e89bec8e07419341098e7f",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "ride_id": {
                                        "type": "integer"
                                    },
                                    "rating": {
                                        "type": "integer"
                                    },
                                    "comment": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Rating submitted"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/subscribe": {
            "post": {
                "tags": [
                    "Subscription"
                ],
                "summary": "Subscribe to the newsletter",
                "description": "Stores an email for newsletter subscription",
                "operationId": "d2480cefcfc2babef16b2bb389cf09bb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "email"
                                ],
                                "properties": {
                                    "email": {
                                        "type": "string",
                                        "example": "user@example.com"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Subscription successful"
                    },
                    "422": {
                        "description": "Validation error"
                    }
                }
            }
        },
...
        "/api/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Get current user's cart",
                "operationId": "0652f94dde92e36bbb3c461356ce1d7b",
                "responses": {
                    "200": {
                        "description": "Cart retrieved successfully"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/add": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Add item to cart",
                "operationId": "408de867ef91c8a1c1298be6461a5e3c",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "item_type",
                                    "item_id",
                                    "quantity"
                                ],
                                "properties": {
                                    "item_type": {
                                        "type": "string",
                                        "example": "product"
                                    },
                                    "item_id": {
                                        "type": "integer"
                                    },
                                    "quantity": {
                                        "type": "integer",
                                        "minimum": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Item added to cart"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/remove/{item_id}": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Remove item from cart",
                "operationId": "384f6db375db64ec88cc644b8a8d023e",
                "parameters": [
                    {
                        "name": "item_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Item removed from cart"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/cart/clear": {
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Clear entire cart",
                "operationId": "c2cf94402630acad1cc6d317ec806693",
                "responses": {
                    "200": {
                        "description": "Cart cleared"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/orders/vendor": {
            "get": {
                "tags": [
                    "Orders"
                ],
                "summary": "Get orders for vendor's products",
                "operationId": "0ae6b9e4a682706af88fd1e56854dbb9",
                "responses": {
                    "200": {
                        "description": "List of vendor's orders"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
        "/api/wallet/transactions": {
            "get": {
                "tags": [
                    "Wallet"
                ],
                "summary": "Get wallet transaction history",
                "operationId": "d7003c8f7d8d61b2b6b5b2d6f2377cc3",
                "responses": {
                    "200": {
                        "description": "List of wallet transactions"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}